Update race_meet_advisor.lua
I tried to keep them rivals, but a bit civil...expect hostile. i'm not good at meme speak, i don't know how to make manipulate be all "hypothetically" and "um, actually", and hostile doesn't seem like a trump to me. Diplomacy is civil and the straight arrow, manipulate is just a bunch of loaded statements, and hostile is just rowdy.