Adding sidequest and modifiers
Mostly finished sidequest and some modifiers. I still want to make some changes to the negotiation itself but I need to find a way to do it without a huge wall of text. And the text needs revision as well.
Based on the time it took me to make this I will try to upload something once a week. I am going to go and work on the propaganda poster sidequest now.