Dialog Updates (#90)
* updated some of the deal's dialogue

I was expecting this to be a short and sweet, bite sized event. Not that i'm complaining about the work, just a bit surprised.

* finished more or less all of the placeholders.

Not all of them are the same quality, but they're all cleaned placeholders at least

* cleaned out the [p] dialogues

gave everything a cleaned dialogue pass, though this is a simple enough event that it isn't a terrible amount of work

* finished dialog

added a commented section for some idea I had, but I later more or less added that to the deny vaccine, asked info block about how the Cult are just giving vaccines to electioneer their voters.

* did most of the suggestions

I only got a bit stumped on the part where you get the intel and the priest has to explain they got the vaccine by taking it. I think my idea for how you refuse the vaccine is just telling the priest that this is just a ploy to get voters and that's what makes the priest reconsider their actions

* Capitalized and changed the intro slightly

Got the capitalized I's in, changed this'll to this would and made the tweak I had in a comment to the first bit

* Update demevent_rebel_arms_deal.lua

* Update demevent_bog_vaccine.lua

* Added possible Revised Desc

Saw the issue for Propaganda Poster, so I wrote a new desc. It's more concise, at least

* Update demevent_collect_tithe.lua

Unfinished but I just needed to save this

* Finished Dialogues

This one has a lot more variety for the vanilla characters than the others, just because I thought of Smith, of course, name dropping his way out of tithes, and then Rook probably blaming an alias for his lack of tithes, and Sal just thinking that she's already been used and abused by the cult enough for one lifetime.

* Gave some of the dialogues more Dem stylings

I made some of the dialogues more in line with the Democratic Race than with Rook's campaign

* Update propaganda_modifier.lua

Co-authored-by: RageLeague <34314733+RageLeague@users.noreply.github.com>